moved isVotedKeyword to keyword_parser.ts
